refactor(tls): Refactor TLS connector structure
